Talkcoin - Restructure changes
Move hard fork variables to util.cpp where they will be made readily
accessible to the rest of the code. Change the developer address
functions in base58.h to always get a height argument and change all
calls to provide that argument. Make building of static binaries default
on Windows.